Ignore psqlrc when running psql commands
PostgreSQL 9.6 seems to change the way psql deals with psqlrc when
running non-interactively. Previously it was silent when parsing the RC
file but it now outputs various messages as it does when running on the
console. This means the postgresql_psql type breaks on PG9.6 if you have
a custom psqlrc installed that causes psql to generate output.

This changeset simply adds the -X option to the command line, telling
psql to ignore any system psqlrc.